Refactored many components into the right folders
