+++ src/components/vehicles/SummaryCard/components/CounterSection.jsx
