+++ src/components/vehicles/SummaryCard/index.jsx
@@ -4,8 +4,8 @@ import {BiDollarCircle, BiTime} from "react-icons/bi";
 import {DateTime} from "luxon";
 import {useMutation} from "@apollo/client";
 import Button from "components/main/Button";
-import SummaryCardSection from "routes/Dashboard/SummaryCard/components/SummaryCardSection";
-import CounterTime from "routes/Dashboard/SummaryCard/components/CounterTime";
+import SummaryCardSection from "components/vehicles/SummaryCard/components/SummaryCardSection";
+import CounterTime from "components/vehicles/SummaryCard/components/CounterTime";
 import DeleModal from "components/modals/DeleModal";
 import CreateCheckInMutation from "mutations/CreateCheckInMutation";
 import CheckedInVehiclesQuery from "queries/CheckedInVehiclesQuery";
